Move ServerTimestamp into its own file
